Added config file and some documentation

+++ anonymise.js
@@ -1,20 +1,19 @@
 const fs = require("fs");
 const path = require("path");
 const readline = require("readline");
-
-// Manulachecking settings
-const SKIP_MANUAL = false;
-let checkedPhrases = [];
+const config = require("./config");
 
 // Diagnostics  
 let nameCounter = 0;
 let manualCounter = 0;
 
+// Open readline if in manual mode
 const rl = readline.createInterface({
     input: process.stdin,
     output: process.stdout,
-  });
-
+});
+// Initialise empty array of already checked phrases
+let checkedPhrases = [];
 
 function manualCheck(phrase, sentence){
     // Function for removing 'phrase' manually¨
@@ -51,6 +50,7 @@ async function checkSentence(sentence){
     const nrCheck = /([0-9]{3}\s[0-9]{2}\s[0-9]{3}|[0-9]{2}\s[0-9]{2}\s[0-9]{2}\s[0-9]{2}|[0-9]{8})/i.test(sentence);
     const phoneCheck = /tlf|telefon|tel/i.test(sentence);
 
+    // Perform trivial checks
     if (contactCheck || emailCheck || nrCheck || phoneCheck) {
         return true;
     }
@@ -59,8 +59,9 @@ async function checkSentence(sentence){
     const nameRegex = /\b[A-ZÆØÅ][a-zæøå]+ [A-ZÆØÅ][a-zæøå]+\b/g;
     const nameCheck = nameRegex.test(sentence);
 
-    nameCounter += Number(nameCheck); 
-    if (SKIP_MANUAL){return false;};
+    nameCounter += Number(nameCheck);
+    if (config.skipManual){return config.nonManualDefault;};
+    // Skip manual categorisation. See config.js.
 
     if (nameCheck) {
         const matches = sentence.match(nameRegex);
@@ -71,7 +72,6 @@ async function checkSentence(sentence){
             if (manCheck) {return true;};
         }
     }
-        
     return false;
 }
 
@@ -89,35 +89,47 @@ const removePersonalInfo  = async () => {
         const adsPath = path.join(__dirname, "ads");
         const employerNames = fs.readdirSync(adsPath);
 
-        // Need to remove status.txt
+        // Need to remove status.txt from ads-folder
         const statusIndex = employerNames.indexOf("status.txt");
         employerNames.splice(statusIndex, 1);
 
         // Loop over all employers
         for (let i = 0; i < employerNames.length; i++){
-            const filename = fs.readdirSync(path.join(adsPath, employerNames[i]));
+            const filename = fs.readdirSync(
+                                path.join(
+                                    adsPath, 
+                                    employerNames[i]
+                                )
+                            );
             
             // Loop over all ad files
             for (let j = 0; j < filename.length; j++){
 
-                let ad = JSON.parse(fs.readFileSync(path.join(adsPath, employerNames[i], filename[j]), (err, data) => {
+                let ad = JSON.parse(
+                    fs.readFileSync(
+                        path.join(
+                            adsPath, 
+                            employerNames[i], 
+                            filename[j]
+                        ), (err, data) => {
                     if (err) throw err;
                     return(data);
                 })); 
 
                 // Remove personal information from ad description
-                let sentences = ad['description'].split(/((?<![A-ZÆØÅ])\.|<p|<h[1-6]|<br|<li)(\s| \/>| >|>)/);
+                let sentences = ad['description']
+                                    .split(/((?<![A-ZÆØÅ])\.|<p|<h[1-6]|<br|<li)(\s| \/>| >|>)/);
                 for (let k = 0; k < sentences.length; k++){
 
                     personalInformation = await checkSentence(sentences[k]);
 
                     if (personalInformation) {
                         // Remove sentence with personal information
-                        sentences[k] = "---PERSONOPPLYSNING---";
+                        sentences[k] = config.anonymityString;
                     }
                 }
                 
-                if (sentences.join() == "---PERSONOPPLYSNING---"){
+                if (sentences.join() == config.anonymityString){
                    // Diagnostic
                     console.log(ad['description']);
                     exit();
@@ -136,7 +148,7 @@ const removePersonalInfo  = async () => {
                         personalInformation = await checkSentence(employerSentences[k]);
     
                         if (personalInformation) {
-                            employerSentences[k] = "---PERSONOPPLYSNING---";
+                            employerSentences[k] = config.anonymityString;
                             cleanedEmployerDescription++;
                         }
                     }
@@ -159,8 +171,8 @@ const removePersonalInfo  = async () => {
             cleanedEmployers++;
         }
 
-        // Close readline after all (potential) manual intervention is completed
-        rl.close();
+        // Close readline after all manual intervention is completed
+        if (!config.skipManual){ rl.close();};
 
         // Add marker in status.txt
         const adStatusPath = path.join(__dirname, "ads", "status.txt");
@@ -171,16 +183,15 @@ const removePersonalInfo  = async () => {
 
         fs.writeFileSync(cleanAdStatusPath, newStatus);
 
-        /*
-        // Remove raw ads
-        fs.rm(adsPath, {recursive : true}, err => {
-            if (err){
-                throw err;
-            }
+        if (config.removeRawAds) {
+            fs.rm(adsPath, {recursive : true}, err => {
+                if (err){
+                    throw err;
+                }
 
-            console.log("Raw ads deleted.")
-        });
-        */
+                console.log("Raw ads deleted.")
+            });
+        }
 
         // Diagnostics
         console.log("Cleaned", cleanedAds, "ads");
@@ -198,5 +209,6 @@ if (!fs.existsSync(path.join(__dirname, "ads"))) {
     console.log("No ads found.");
 } else {
     console.log("Ad folder found.");
+
     removePersonalInfo();
 }
